refactor(cockpit): extract complexity gate module
